+++ src/components/tmCountdown/index.js
@@ -1,0 +1,5 @@
+import tmCountdown from "./tmCountdown.vue"
+
+export default function install(Vue) {
+  Vue.component("tm-countdown", tmCountdown)
+}