+++ CHANGELOG.md
@@ -12,7 +12,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added `./src/index.js` to support `import`
 - Added a README and CHANGELOG
 - Added a placeholder HelloWorld component
+<<<<<<< HEAD
 - Added @nylira/vue-notifications as `tmNotifications`
+=======
+- Added @nylira/vue-countdown as `tmCountdown`
+>>>>>>> 6c467c99375f0f692bbaee1a32baf8f6928e8896
 
 ### Changes